refactor(perp): update order functions for binance

+++ exchange/perp.go
@@ -55,7 +55,7 @@ type PerpExchange interface {
 	// ========== 合约特有功能 ==========
 
 	// SetLeverage 设置杠杆
-	SetLeverage(ctx context.Context, symbol string, leverage int) error
+	SetLeverage(ctx context.Context, symbol string, leverage int64) error
 
 	// SetMarginType 设置保证金类型（isolated/cross）
 	SetMarginType(ctx context.Context, symbol string, marginType option.MarginType) error
